Develop (#29)
* Added aws-lambda-powertools

* Added powertools logger

* Format with black

* Updated Pipfile.lock

* Added certifi package

* Revert "Added certifi package"

This reverts commit 2b89981dccf4169b264b64603173945e5dab55bc.

* Disabled pipenv caching for debugging (from yesterday to today the workflow is broke by itself 🤷)

* There is no 'cache: false'

* Revert "There is no 'cache: false'"

This reverts commit 508432114d630c47104f1cc17b00d12d36d94900.

* Revert "Disabled pipenv caching for debugging (from yesterday to today the workflow is broke by itself 🤷)"

This reverts commit 206a91103e8928cacd465fe292ff3306efd3b068.

* Try ubuntu-22.04

* Revert "Try ubuntu-22.04"

This reverts commit 38b576b1a8bf05902fce873b5afa96d7c4777de2.

* List installed pipenv packages

* Display venv path after install

* Revert "Display venv path after install"

This reverts commit 9e097fbbc2836dcbc20c5c4e6c737980ea9dad97.

* Install certifi package with pipenv

* Revert "Install certifi package with pipenv"

This reverts commit 7ec0a5b523b6a3526f7df73d51769401dd8aba63.

* Added aws-lambda-powertools package

* Revert "Added aws-lambda-powertools package"

This reverts commit 8b5fc6c65b65e08a4927525c0b02626d0d7ee3da.

* Added aws-lambda-powertools package (1.26.7)

* Revert "Added aws-lambda-powertools package (1.26.7)"

This reverts commit a15a4a2140dba320e751e067da6623c40fa2761c.

* Added aws-lambda-powertools

* Added certifi

* Format with black

* Skip app tests until fix

* Revert "Skip app tests until fix"

This reverts commit 3ce90f9437736b99cb4ff605a09903b635d4b403.

* Removed pytest-httpx for testing

* Replaced pytest-httpx with respx

* Refactor

* Removed autopep8

* Refactor

* Reformat with black

* Refactor

* Added aws-lambda-powertools logger, metrics and tracer

* Replaced logging with aws-lambda-powertools logger

* Revert "Replaced logging with aws-lambda-powertools logger"

This reverts commit ed7d33de185b000cc3583dc4d491915864286bcc.

* Fixed typo

* Removed utils.py

* Replaced logging with aws_lambda_powertools logger

* Added post repository

* Refactor

* Added post repository tests

* Updated serverless-offline version from 8.8.0 to 9.2.4

* Removed serverless-offline related configuration

* Added faulty tests

* Added faulty tests

* Raise exception based on auto_error

* Refactor

* Simplified error details

* Added token to routes and request body validation set to manual

* Added filter expression as parameter

+++ app/api/v1/routes/posts.py
@@ -1,38 +1,46 @@
 from typing import List
 
 from aws_lambda_powertools.metrics import Metrics, MetricUnit
-from fastapi import APIRouter, status, Depends
+from fastapi import status, APIRouter, Depends, Request, HTTPException
 from starlette.responses import Response
 
-from app.auth import JWTBearer
+from app.auth import JWTBearer, JWTToken
 from app.models.post import Post
 from app.schemas.post import CreatePost, UpdatePost
 from app.services.post import PostService
 
-jwt_bearer = JWTBearer()
+jwt_bearer = JWTBearer(auto_error=False)
 metrics = Metrics()
 post_service = PostService()
 router = APIRouter()
 
 
-@router.post('', dependencies=[Depends(jwt_bearer)])
-async def create_post(body: CreatePost) -> Response:
-    post = await post_service.create_post(body.dict())
-    metrics.add_metric(name='CreatePost', unit=MetricUnit.Count, value=1)
-    return Response(
-        status_code=status.HTTP_201_CREATED,
-        headers={'Location': f'/api/v1/posts/{post.id}'},
-    )
+@router.post('')
+async def create_post(
+    request: Request, token: JWTToken = Depends(jwt_bearer)
+) -> Response:
+    if token:
+        model = CreatePost.parse_raw(await request.body())
+        post = await post_service.create_post(model.dict())
+        metrics.add_metric(name='CreatePost', unit=MetricUnit.Count, value=1)
+        return Response(
+            status_code=status.HTTP_201_CREATED,
+            headers={'Location': f'/api/v1/posts/{post.id}'},
+        )
+    else:
+        raise await create_http_exception()
 
 
 @router.delete(
     '/{uuid}',
-    dependencies=[Depends(jwt_bearer)],
     status_code=status.HTTP_204_NO_CONTENT,
 )
-async def delete_post(uuid: str):
-    await post_service.delete_post(uuid)
-    metrics.add_metric(name='DeletePost', unit=MetricUnit.Count, value=1)
+async def delete_post(uuid: str, token: JWTToken = Depends(jwt_bearer)):
+    if token:
+        await post_service.delete_post(uuid)
+        metrics.add_metric(name='DeletePost', unit=MetricUnit.Count, value=1)
+    else:
+        raise await create_http_exception()
 
 
 @router.get('', status_code=status.HTTP_200_OK)
@@ -51,9 +59,20 @@ async def get_post_by_uuid(uuid: str) -> Post:
 
 @router.put(
     '/{uuid}',
-    dependencies=[Depends(jwt_bearer)],
     status_code=status.HTTP_204_NO_CONTENT,
 )
-async def update_post(uuid: str, data: UpdatePost):
-    await post_service.update_post(uuid, data.dict())
-    metrics.add_metric(name='UpdatePost', unit=MetricUnit.Count, value=1)
+async def update_post(
+    request: Request, uuid: str, token: JWTToken = Depends(jwt_bearer)
+):
+    if token:
+        model = UpdatePost.parse_raw(await request.body())
+        await post_service.update_post(uuid, model.dict())
+        metrics.add_metric(name='UpdatePost', unit=MetricUnit.Count, value=1)
+    else:
+        raise await create_http_exception()
+
+
+async def create_http_exception(
+    status_code: int = status.HTTP_403_FORBIDDEN, detail: str = 'Not authenticated'
+):
+    raise HTTPException(status_code=status_code, detail=detail)
